requirements.txt
pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
xgboost>=1.6
mlflow>=2.3
azure-ai-ml>=1.11
azure-identity>=1.14
pytest>=7.0
